moving database package to data layer
