+++ app/tooling/sales-admin/commands/seed.go
@@ -5,13 +5,13 @@ import (
 	"fmt"
 	"time"
 
+	db "github.com/ardanlabs/service/business/data/database/pgx"
 	"github.com/ardanlabs/service/business/data/dbmigrate"
-	database "github.com/ardanlabs/service/business/sys/database/pgx"
 )
 
 // Seed loads test data into the database.
-func Seed(cfg database.Config) error {
-	db, err := database.Open(cfg)
+func Seed(cfg db.Config) error {
+	db, err := db.Open(cfg)
 	if err != nil {
 		return fmt.Errorf("connect database: %w", err)
 	}
