+++ app/tooling/sales-admin/commands/useradd.go
@@ -9,18 +9,18 @@ import (
 	"github.com/ardanlabs/service/business/core/event"
 	"github.com/ardanlabs/service/business/core/user"
 	"github.com/ardanlabs/service/business/core/user/stores/userdb"
-	database "github.com/ardanlabs/service/business/sys/database/pgx"
+	db "github.com/ardanlabs/service/business/data/database/pgx"
 	"github.com/ardanlabs/service/business/sys/logger"
 )
 
 // UserAdd adds new users into the database.
-func UserAdd(log *logger.Logger, cfg database.Config, name, email, password string) error {
+func UserAdd(log *logger.Logger, cfg db.Config, name, email, password string) error {
 	if name == "" || email == "" || password == "" {
 		fmt.Println("help: useradd <name> <email> <password>")
 		return ErrHelp
 	}
 
-	db, err := database.Open(cfg)
+	db, err := db.Open(cfg)
 	if err != nil {
 		return fmt.Errorf("connect database: %w", err)
 	}
