+++ business/core/user/stores/userdb/userdb.go
@@ -11,10 +11,10 @@ import (
 	"net/mail"
 
 	"github.com/ardanlabs/service/business/core/user"
+	"github.com/ardanlabs/service/business/data/database"
+	db "github.com/ardanlabs/service/business/data/database/pgx"
+	"github.com/ardanlabs/service/business/data/database/pgx/dbarray"
 	"github.com/ardanlabs/service/business/data/order"
-	"github.com/ardanlabs/service/business/sys/database"
-	db "github.com/ardanlabs/service/business/sys/database/pgx"
-	"github.com/ardanlabs/service/business/sys/database/pgx/dbarray"
 	"github.com/ardanlabs/service/business/sys/logger"
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
